# Match Torch speed for sum reduction on M1