--- a/tinygrad/lazy.py
+++ b/tinygrad/lazy.py
@@ -210,6 +210,12 @@ class LazyBuffer:
   def reduce_op(self:LazyBuffer, op:ReduceOps, new_shape:Tuple[int, ...]) -> LazyBuffer:
     if self.shape == tuple(new_shape): return self
     srcs = _push_movement_ops((self,)) if SHUFFLE_MOVEMENT_OPS else (self,)
+    
+    differences = [a != b for a,b in zip(self.shape, new_shape)]
+    if sum(differences) > 1:
+      intermediate_shape = tuple(new if i == differences.index(True) else old for i, (new, old) in enumerate(zip(new_shape, self.shape)))
+      return create_lazybuffer(self.device, ShapeTracker(intermediate_shape), ReduceOps, LazyOp(op, srcs, intermediate_shape), self.dtype).reduce_op(op, new_shape)
+
     return create_lazybuffer(self.device, ShapeTracker(new_shape), ReduceOps, LazyOp(op, srcs, new_shape), self.dtype)
 
   def reshape(self:LazyBuffer, arg:Tuple[int, ...]) -> LazyBuffer: