@tinyb0t 2023-07-08
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/lazy.py                           7               0
------------------------------------------------------------
lines added in the tinygrad folder: 7
```