@geohot 2023-07-08
Bounty locked!

Re: two kernels, do we have a final answer on if PyTorch is doing this?