@nimlgen 2023-07-08
Working on reduces also got some major speed up on these tests (https://github.com/tinygrad/tinygrad/pull/1137). I decided to go with atomics to speed up slow reduces. I just got 80x slower to 8x slower, but cannot test this on m1.

The problem it worked slow, since it was generated with only 1 block (so only 1 SM is busy?). So the way here is to use atomics. I am able to generate kernel like this in my POC:
```
#include <metal_stdlib>
using namespace metal;
kernel void r1_1024_256_64(device atomic_float* data0, const device float* data1, uint3 gid [[threadgroup_position_in_grid]], uint3 lid [[thread_position_in_threadgroup]]) {
threadgroup float temp[256];
{ int gidx0 = gid.x;  /* 1024 */
  { int lidx1 = lid.x;  /* 256 */
    float acc_0_0 = 0.0f;
    for (int ridx2 = 0; ridx2 <= 63; ++ridx2) {
      float val_1_0 = data1[((gidx0*64)+(lidx1*65536)+ridx2)];
      acc_0_0 = (val_1_0+acc_0_0);
    } /* reduce */
    temp[lidx1] = acc_0_0;
    threadgroup_barrier(mem_flags::mem_threadgroup);
    if (lidx1 == 0) {
    float acc_m1_0 = 0.0f;
    {
    for (int tidx1 = 0; tidx1 <= 255; ++tidx1) {
      float val_m1_0 = temp[tidx1];
      acc_m1_0 = (val_m1_0+acc_m1_0);
    }} /* late_reduce */
    atomic_fetch_add_explicit(data0, acc_m1_0, memory_order_relaxed);
  }} /* local */
} /* global */
}
```

@geohot what's your opinion to add atomic_sum/max to backends (maybe as optional for be to implement)?