@geohot 2023-07-08
I don't understand all the implications of atomics.
* Is torch using them?
* Are they in OpenCL? What version?
* What would it look like as a change? Maybe it's a flag on the store?