@nimlgen 2023-07-08
The main problem with generated reduces is that global_dim=1 and local_dim=256 (from the test_sum). It is just 256 threads (8 warps) running on 1 SM. This simply can't saturate the whole memory bandwidth. [The internet says](https://news.ycombinator.com/item?id=25420290#:~:text=in%20Software%20...-,The%20M1%20has%20pretty%20high%20memory%20latency%20at%20around%20100,or%20Intel%20for%20typical%20systems) m1 main memory has a latency of 100ns. Assuming each of 256 threads load 4bytes, we got 1kb in flight every 100ns. This seems to be around 10gb/s. The only way to solve this is to have more threads to hide latency, the way to do this is to increase the global dim. This requires atomics, since we are going to write data in parallel manner (from global dim).

Apple has the similar implementation of reduce as example in their [metal specs](metal/Metal-Shading-Language-Specification) (p.179). This makes me believe that pytorch might use the same approach.

Currently I have support for atomics for metal only. Not sure if OpenCL supports atomics on floats but there are [some workarounds](https://streamhpc.com/blog/2016-02-09/atomic-operations-for-floats-in-opencl-improved/). CUDA has support for float atomics and the same for 7900 xtx which [has VK_EXT_shader_atomic_float](https://vulkan.gpuinfo.org/displayreport.php?id=22198#features_extensions)).

Yeah, flags for stores sounds great.