@madebyollin 2023-07-08
I was curious about "what is PyTorch MPS doing for reduces"
1. From reading source, PyTorch just [calls](https://github.com/pytorch/pytorch/blob/5da4745c24859d501c9555dbe67a20dfcd199942/aten/src/ATen/native/mps/operations/ReduceOps.mm#L229C38-L229C60) MPSGraph's [`reductionSumWithTensor`](https://developer.apple.com/documentation/metalperformanceshadersgraph/mpsgraph/3580494-reductionsumwithtensor?language=objc).
2. From Instruments profiling, MPSGraph `reductionSumWithTensor` uses a compute shader `reduce_multiple_passes_axes_add` multiple times per single reduce:
<img width="1322" alt="image" src="https://github.com/tinygrad/tinygrad/assets/8733615/6e8a4072-1abd-4ac4-b9dd-edd9274203dc">
3. From grepping, it looks like this shader is part of the MPSNDArray framework:

```bash
/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/metal-objdump \
-disassemble-all \
/System/Library/Frameworks/MetalPerformanceShaders.framework/Versions/A/Frameworks/MPSNDArray.framework/Versions/Current/Resources/default.metallib \
| sed -n '/reduce_multiple_passes_axes_add/,/reduce_multiple_passes_axes_min/p'
```