@geohot 2023-07-09
I think this should apply to all GPU-like backends, OpenCL, metal, cuda. We also have a lot more test coverage that way.