@ae99 2023-07-09
@nimlgen with this change, the primary generated kernel uses global_dim=128 and localdim=32, for example see one of the generated kernels below:
```c++
kernel void r_128_32_1024_4(device float* data0, const device float* data1, uint3 gid [[threadgroup_position_in_grid]], uint3 lid [[thread_position_in_threadgroup]]) {
{ int gidx0 = gid.x;  /* 128 */
  { int lidx1 = lid.x;  /* 32 */
    float acc0_0 = 0.0f;
    for (int ridx2 = 0; ridx2 <= 1023; ++ridx2) {
      float4 val1_0 = (float4)(*((device float4*)(data1+((gidx0*131072)+(lidx1*4096)+(ridx2*4)))));
      acc0_0 = (val1_0.x+acc0_0);
      acc0_0 = (val1_0.y+acc0_0);
      acc0_0 = (val1_0.z+acc0_0);
      acc0_0 = (val1_0.w+acc0_0);
    } /* reduce */
    data0[((gidx0*32)+lidx1)] = acc0_0;
  }} /* global+local */
}
```
So for single kernel approach you are spot on that this is the bottleneck, but for two-kernel we're actually hitting on all the major optimizations you mention. 

To my knowledge metal doesn't support atomics for floats, only int types, so there doesn't seem to be a workable solution using them with that path but correct me if that's changed.

Thanks @madebyollin, thought this must be the case but good to see confirmed!