@ae99 2023-07-09
@geohot I've updated the PR to be a bit cleaner. I've checked and it appears to me that existing `test/test_ops.py` test cases cover off correctness of this optimization quite well. I'd say might be ready for a first review?


> I think this should apply to all GPU-like backends, OpenCL, metal, cuda. We also have a lot more test coverage that way.

Is this in reference to `self.device == "METAL"` in this PR or referencing atomics discussion? 


Side note: we can achieve approx 250 GB/s with a larger matrix and a few other tweaks around the codebase, but they all start clashing with other optimizations, so current change is the best I've gotton for 4096x4096 and smaller matrices.