@geohot 2023-07-09
I want two kernels on all GPU-like backends, nothing about atomics. That will also trigger all the tests.